Factor out `_1DApp` and rename it to be `OneDimBaseApp`
